% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPoisson.R
\name{gPoisson}
\alias{gPoisson}
\title{Computes g(z,t) using the theta2 Poisson series (27)}
\usage{
gPoisson(x, t, Nmax = 5)
}
\arguments{
\item{Nmax, }{the truncation order of the series. Defaults to 5.}
}
\description{
Uses the Poisson series (27)
}
\examples{
gPoisson(0.0,1.0) should return 0.901925
}
