% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRec.R
\name{gRec}
\alias{gRec}
\title{Computes the g(z,t) function by recursion}
\usage{
gRec(x, t, Nmax = 5)
}
\arguments{
\item{Nmax}{is the truncation order in the evaluation of the Appell sums. Defaults to 5}
}
\description{
The function g(z,t) is computed by recursion from its values in the primitive cell (-1/2 t,1/2 t)
}
\examples{
gRec(0.0,1.0) 
}
