% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiLNexact.R
\name{phiLNexact}
\alias{phiLNexact}
\title{exact evaluation of phi(k*t,t) with integer k, using (20) in JCAM}
\usage{
phiLNexact(t, k)
}
\arguments{
\item{t}{}

\item{k}{}
}
\description{
exact evaluation of phi(k*t,t) with integer k, using (20) in JCAM
}
