% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiLogisticNormal.R
\name{phiLogisticNormal}
\alias{phiLogisticNormal}
\title{Logistic normal integral}
\usage{
phiLogisticNormal(z, t, Nmax = 5)
}
\arguments{
\item{z}{}

\item{t}{}

\item{Nmax}{}
}
\value{
Returns a real value
}
\description{
Logistic normal integral
}
